pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.8",
    "greenlet==2.0.2",
    "haystack-ai==2.2.0",
    "json-repair>=0.39.1",
    "pydantic>=2.10.6",
    "pypdf>=5.3.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]
